requirements.txt
sentence-transformers
pinecone-client 
requests
pandas
Flask
Flask_Cors
Flask_RESTful
pymongo
